requirements.txt
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
Werkzeug==1.0.1
autopep8==2.0.2
colorama==0.4.6
importlib-metadata==6.1.0
numpy==1.24.2
Pillow==9.4.0
pycodestyle==2.10.0
tomli==2.0.1
zipp==3.15.0
keras
tensorflow
replicate
python-dotenv
